backend/requirements.txt
Flask==3.0.3
flask-pymongo==2.3.0
flask-cors==4.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
authlib==1.3.1
requests==2.32.3
Werkzeug==3.0.3
opencv-python==4.9.0.80
